Chore: 파일 구조 생성 (#1)
